fix: fixed reordering issues in dyn system
Fixed the reodering issue in the dynamic update system which was caused
by the use of `pack_end()` for adding a child to box_container in the create_widget() fn.

`pack_end()` is a horrible option as it would require hardcoding
"expanded", "fill" etc.

This fix uses the better option `add()` which fixes all the reodering
issues.

According to all the tests, (did 15+ tests in a dyn fn that would return 3
widgets) non of them had ordering issues. So, I will mark this as a fix.