feat: better errors with rhai_trace (lib built for ewwii)
With this commit, Ewwii now integrates rhai_trace, which is a rust library that I
built to generate better errors with rhai.

It is plugged into codespan reporting for providing awesome errors which
makes programming with ewwii much easier! It is a major UX update along
with the removal of `dyn_id` dependance.